+++ src/components/highlight/css.js
@@ -1,0 +1,7 @@
+import Prism from 'prismjs';
+
+function highlightCss(rawCode) {
+  return rawCode ? Prism.highlight(rawCode, Prism.languages['css']) : '';
+}
+
+export { highlightCss };